:recycle: refactor : remove or rename enums package
