+++ src/main/java/com/run_us/server/domains/running/exception/RunningNotFoundException.java
@@ -1,7 +1,7 @@
 package com.run_us.server.domains.running.exception;
 
 import com.run_us.server.global.exception.BusinessException;
-import com.run_us.server.global.exception.enums.CustomResponseCode;
+import com.run_us.server.global.exception.code.CustomResponseCode;
 
 public class RunningNotFoundException extends BusinessException {
